% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae.R
\docType{data}
\name{adae}
\alias{adae}
\title{Adverse Events Analysis}
\format{
A data frame with 107 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Time of Last Exposure Imput. Flag}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHDTF }{Date of Death Imputation Flag}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ AESEQ }{Sequence Number}
\item{ AETERM }{Reported Term for the Adverse Event}
\item{ AEDECOD }{Dictionary-Derived Term}
\item{ AEBODSYS }{Body System or Organ Class}
\item{ AEBDSYCD }{Body System or Organ Class Code}
\item{ AELLT }{Lowest Level Term}
\item{ AELLTCD }{Lowest Level Term Code}
\item{ AEPTCD }{Preferred Term Code}
\item{ AEHLT }{High Level Term}
\item{ AEHLTCD }{High Level Term Code}
\item{ AEHLGT }{High Level Group Term}
\item{ AEHLGTCD }{High Level Group Term Code}
\item{ AESOC }{Primary System Organ Class}
\item{ AESOCCD }{Primary System Organ Class Code}
\item{ AESTDTC }{Start Date/Time of Adverse Event}
\item{ ASTDT }{Analysis Start Date}
\item{ ASTDTM }{Analysis Start Date/Time}
\item{ ASTDTF }{Analysis Start Date Imputation Flag}
\item{ ASTTMF }{Analysis Start Time Imputation Flag}
\item{ AEENDTC }{End Date/Time of Adverse Event}
\item{ AENDT }{Analysis End Date}
\item{ AENDTM }{Analysis End Date/Time}
\item{ AENDTF }{Analysis End Date Imputation Flag}
\item{ AENTMF }{Analysis End Time Imputation Flag}
\item{ ASTDY }{Analysis Start Relative Day}
\item{ AESTDY }{Study Day of Start of Adverse Event}
\item{ AENDY }{Analysis End Relative Day}
\item{ AEENDY }{Study Day of End of Adverse Event}
\item{ ADURN }{Analysis Duration (N)}
\item{ ADURU }{Analysis Duration Units}
\item{ TRTEMFL }{Treatment Emergent Analysis Flag}
\item{ AOCCIFL }{1st Max Sev./Int. Occurrence Flag}
\item{ AESER }{Serious Event}
\item{ AESDTH }{Results in Death}
\item{ AESLIFE }{Is Life Threatening}
\item{ AESHOSP }{Requires or Prolongs Hospitalization}
\item{ AESDISAB }{Persist or Signif Disability/Incapacity}
\item{ AESCONG }{Congenital Anomaly or Birth Defect}
\item{ AESEV }{Severity/Intensity}
\item{ ASEV }{Analysis Severity/Intensity}
\item{ ASEVN }{Analysis Severity/Intensity (N)}
\item{ AEREL }{Causality}
\item{ AREL }{Analysis Causality}
\item{ AEACN }{Action Taken with Study Treatment}
\item{ AESPID }{Sponsor-Defined Identifier}
\item{ AEOUT }{Outcome of Adverse Event}
\item{ AESCAN }{Involves Cancer}
\item{ AESOD }{Occurred with Overdose}
\item{ AEDTC }{Date/Time of Collection}
\item{ LDOSEDTM }{End Date/Time of Last Dose}
\item{ DOSEON }{Treatment Dose}
\item{ DOSEU }{Treatment Dose Unit}
}
}
\source{
Generated from admiral package (template ad_adae.R).
}
\usage{
adae
}
\description{
Adverse Events Analysis
}
\examples{
data("adae")
}
\references{
None
}
\keyword{datasets}
