% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcva_ophtha.R
\docType{data}
\name{adbcva_ophtha}
\alias{adbcva_ophtha}
\title{Dataset adbcva_ophtha}
\format{
A data frame with 116 columns:
\describe{
\item{ STUDYID }{undocumented field}
\item{ DOMAIN }{undocumented field}
\item{ USUBJID }{undocumented field}
\item{ OESEQ }{undocumented field}
\item{ OECAT }{undocumented field}
\item{ OESCAT }{undocumented field}
\item{ OEDTC }{undocumented field}
\item{ VISIT }{undocumented field}
\item{ VISITNUM }{undocumented field}
\item{ VISITDY }{undocumented field}
\item{ OESTRESN }{undocumented field}
\item{ OESTRESC }{undocumented field}
\item{ OEORRES }{undocumented field}
\item{ OETEST }{undocumented field}
\item{ OETESTCD }{undocumented field}
\item{ OETSTDTL }{undocumented field}
\item{ OELAT }{undocumented field}
\item{ OELOC }{undocumented field}
\item{ OEDY }{undocumented field}
\item{ OEMETHOD }{undocumented field}
\item{ OEORRESU }{undocumented field}
\item{ OESTRESU }{undocumented field}
\item{ OESTAT }{undocumented field}
\item{ OETPT }{undocumented field}
\item{ OETPTNUM }{undocumented field}
\item{ TRTSDT }{undocumented field}
\item{ TRTEDT }{undocumented field}
\item{ TRT01A }{undocumented field}
\item{ TRT01P }{undocumented field}
\item{ STUDYEYE }{undocumented field}
\item{ AVAL }{undocumented field}
\item{ AVALU }{undocumented field}
\item{ DTYPE }{undocumented field}
\item{ AFEYE }{undocumented field}
\item{ PARAM }{undocumented field}
\item{ PARAMCD }{undocumented field}
\item{ AVALC }{undocumented field}
\item{ ADT }{undocumented field}
\item{ ADY }{undocumented field}
\item{ ATPTN }{undocumented field}
\item{ ATPT }{undocumented field}
\item{ AVISIT }{undocumented field}
\item{ AVISITN }{undocumented field}
\item{ BASETYPE }{undocumented field}
\item{ ONTRTFL }{undocumented field}
\item{ ABLFL }{undocumented field}
\item{ ANL01FL }{undocumented field}
\item{ ANL02FL }{undocumented field}
\item{ WORS01FL }{undocumented field}
\item{ BASE }{undocumented field}
\item{ BASEC }{undocumented field}
\item{ CHG }{undocumented field}
\item{ PCHG }{undocumented field}
\item{ ASEQ }{undocumented field}
\item{ SUBJID }{undocumented field}
\item{ RFSTDTC }{undocumented field}
\item{ RFENDTC }{undocumented field}
\item{ RFXSTDTC }{undocumented field}
\item{ RFXENDTC }{undocumented field}
\item{ RFICDTC }{undocumented field}
\item{ RFPENDTC }{undocumented field}
\item{ DTHDTC }{undocumented field}
\item{ DTHFL }{undocumented field}
\item{ SITEID }{undocumented field}
\item{ AGE }{undocumented field}
\item{ AGEU }{undocumented field}
\item{ SEX }{undocumented field}
\item{ RACE }{undocumented field}
\item{ ETHNIC }{undocumented field}
\item{ ARMCD }{undocumented field}
\item{ ARM }{undocumented field}
\item{ ACTARMCD }{undocumented field}
\item{ ACTARM }{undocumented field}
\item{ COUNTRY }{undocumented field}
\item{ DMDTC }{undocumented field}
\item{ DMDY }{undocumented field}
\item{ TRTSDTM }{undocumented field}
\item{ TRTSTMF }{undocumented field}
\item{ TRTEDTM }{undocumented field}
\item{ TRTETMF }{undocumented field}
\item{ TRTDURD }{undocumented field}
\item{ SCRFDT }{undocumented field}
\item{ EOSDT }{undocumented field}
\item{ EOSSTT }{undocumented field}
\item{ FRVDT }{undocumented field}
\item{ RANDDT }{undocumented field}
\item{ DTHDT }{undocumented field}
\item{ DTHADY }{undocumented field}
\item{ LDDTHELD }{undocumented field}
\item{ LSTALVDT }{undocumented field}
\item{ AGEGR1 }{undocumented field}
\item{ SAFFL }{undocumented field}
\item{ RACEGR1 }{undocumented field}
\item{ REGION1 }{undocumented field}
\item{ LDDTHGR1 }{undocumented field}
\item{ DTH30FL }{undocumented field}
\item{ DTHA30FL }{undocumented field}
\item{ DTHB30FL }{undocumented field}
\item{ CRIT1 }{undocumented field}
\item{ CRIT1FL }{undocumented field}
\item{ CRIT2 }{undocumented field}
\item{ CRIT2FL }{undocumented field}
\item{ CRIT3 }{undocumented field}
\item{ CRIT3FL }{undocumented field}
\item{ CRIT4 }{undocumented field}
\item{ CRIT4FL }{undocumented field}
\item{ CRIT5 }{undocumented field}
\item{ CRIT5FL }{undocumented field}
\item{ CRIT6 }{undocumented field}
\item{ CRIT6FL }{undocumented field}
\item{ CRIT7 }{undocumented field}
\item{ CRIT7FL }{undocumented field}
\item{ CRIT8 }{undocumented field}
\item{ CRIT8FL }{undocumented field}
\item{ AVALCA1N }{undocumented field}
\item{ AVALCAT1 }{undocumented field}
}
}
\source{
Generated from admiralophtha package (template ad_adbcva.R).
}
\usage{
adbcva_ophtha
}
\description{
adbcva_ophtha dataset
}
\examples{
data("adbcva_ophtha")
}
\references{
None
}
\keyword{datasets}
