% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adce_vaccine.R
\docType{data}
\name{adce_vaccine}
\alias{adce_vaccine}
\title{Clinical Events Analysis for Vaccine}
\format{
A data frame with 56 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ ASEQ }{Analysis Sequence Number}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ ETHNIC }{Ethnicity}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ APERSDT }{Period Start Date}
\item{ APEREDT }{Period End Date}
\item{ APERSTDY }{Analysis Sub-period Start Relative Day}
\item{ CESEQ }{Sequence Number}
\item{ CETERM }{Reported Term for the Clinical Event}
\item{ CEDECOD }{Dictionary-Derived Term}
\item{ CECAT }{Category for the Clinical Event}
\item{ CESCAT }{Subcategory for the Clinical Event}
\item{ CESTDTC }{Start Date/Time of Clinical Event}
\item{ ASTDT }{Analysis Start Date}
\item{ CEENDTC }{End Date/Time of Clinical Event}
\item{ AENDT }{Analysis End Date}
\item{ ASTDY }{Analysis Start Relative Day}
\item{ AENDY }{Analysis End Relative Day}
\item{ ADURN }{Analysis Duration (N)}
\item{ ADURU }{Analysis Duration Units}
\item{ CEDUR }{Duration of Clinical Event}
\item{ APERIOD }{Period}
\item{ CEOCCUR }{Clinical Event Occurrence}
\item{ CEPRESP }{Clinical Event Pre-specified}
\item{ AOCC01FL }{Event Occurrence Flag}
\item{ ASEV }{Analysis Severity/Intensity}
\item{ ASEVN }{Analysis Severity/Intensity (N)}
\item{ AREL }{Analysis Causality}
\item{ CELNKID }{Link ID}
\item{ CELNKGRP }{Link Group ID}
\item{ CELAT }{Laterality}
\item{ CELOC }{Location of Event}
\item{ CESEV }{Severity/Intensity}
\item{ CEREL }{Causality}
\item{ CEOUT }{Outcome of Event}
\item{ EPOCH }{Epoch}
\item{ CEDTC }{Date/Time of Event Collection}
\item{ CETPT }{Planned Time Point Name}
\item{ CETPTNUM }{Planned Time Point Number}
\item{ CETPTREF }{Time Point Reference}
\item{ CERFTDTC }{Date/Time of Reference Time Point}
\item{ CEEVINTX }{Evaluation Interval Text}
\item{ CESTAT }{Completion Status}
\item{ CEREASND }{Reason Clinical Event Not Collected}
}
}
\source{
Generated from admiralvaccine package (template ad_adce.R).
}
\usage{
adce_vaccine
}
\description{
Clinical Events Analysis for Vaccine
}
\examples{
data("adce_vaccine")
}
\references{
None
}
\keyword{dataset}
\keyword{vaccine}
