% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\docType{data}
\name{adcp}
\alias{adcp}
\title{Dataset adcp}
\format{
A data frame with 122 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ NFRLT }{NFRLT}
\item{ PCTESTCD }{PCTESTCD}
\item{ PCTEST }{PCTEST}
\item{ PCORRES }{PCORRES}
\item{ PCORRESU }{PCORRESU}
\item{ PCSTRESC }{PCSTRESC}
\item{ PCSTRESN }{PCSTRESN}
\item{ PCSTRESU }{PCSTRESU}
\item{ PCNAM }{PCNAM}
\item{ PCSPEC }{PCSPEC}
\item{ PCLLOQ }{PCLLOQ}
\item{ VISIT }{VISIT}
\item{ VISITNUM }{VISITNUM}
\item{ PCDTC }{PCDTC}
\item{ PCDY }{PCDY}
\item{ PCTPT }{PCTPT}
\item{ PCTPTNUM }{PCTPTNUM}
\item{ TRTSDT }{TRTSDT}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRT01P }{TRT01P}
\item{ TRT01A }{TRT01A}
\item{ ADTM }{ADTM}
\item{ ATMF }{ATMF}
\item{ ADT }{ADT}
\item{ ATM }{ATM}
\item{ ADY }{ADY}
\item{ FANLDTM }{FANLDTM}
\item{ AVISITN }{AVISITN}
\item{ AVISIT }{AVISIT}
\item{ ASTDT }{ASTDT}
\item{ ASTDTM }{ASTDTM}
\item{ AENDT }{AENDT}
\item{ AENDTM }{AENDTM}
\item{ ASTTM }{ASTTM}
\item{ AENTM }{AENTM}
\item{ AFRLT }{AFRLT}
\item{ ARRLT }{ARRLT}
\item{ PCRFTDTM }{PCRFTDTM}
\item{ FANLDT }{FANLDT}
\item{ FANLTM }{FANLTM}
\item{ PCRFTDT }{PCRFTDT}
\item{ PCRFTTM }{PCRFTTM}
\item{ NRRLT }{NRRLT}
\item{ ATPTN }{ATPTN}
\item{ ATPT }{ATPT}
\item{ ATPTREF }{ATPTREF}
\item{ ABLFL }{ABLFL}
\item{ BASETYPE }{BASETYPE}
\item{ DOSEA }{DOSEA}
\item{ DOSEP }{DOSEP}
\item{ DOSEU }{DOSEU}
\item{ FRLTU }{FRLTU}
\item{ RRLTU }{RRLTU}
\item{ PARAMCD }{PARAMCD}
\item{ ALLOQ }{ALLOQ}
\item{ AVAL }{AVAL}
\item{ AVALU }{AVALU}
\item{ AVALCAT1 }{AVALCAT1}
\item{ SRCDOM }{SRCDOM}
\item{ SRCVAR }{SRCVAR}
\item{ SRCSEQ }{SRCSEQ}
\item{ DTYPE }{DTYPE}
\item{ MRRLT }{MRRLT}
\item{ ANL01FL }{ANL01FL}
\item{ ANL02FL }{ANL02FL}
\item{ BASE }{BASE}
\item{ CHG }{CHG}
\item{ ASEQ }{ASEQ}
\item{ PARAM }{PARAM}
\item{ PARAMN }{PARAMN}
\item{ HTBL }{HTBL}
\item{ HTBLU }{HTBLU}
\item{ WTBL }{WTBL}
\item{ WTBLU }{WTBLU}
\item{ BMIBL }{BMIBL}
\item{ BMIBLU }{BMIBLU}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTEDT }{TRTEDT}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adcp
}
\description{
adcp dataset
}
\examples{
data("adcp")
}
\references{
None
}
\keyword{datasets}
