% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{Laboratory Analysis}
\format{
A data frame with 115 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ ASEQ }{Analysis Sequence Number}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRTP }{Planned Treatment}
\item{ TRTA }{Actual Treatment}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Time of Last Exposure Imput. Flag}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHDTF }{Date of Death Imputation Flag}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ AVISIT }{Analysis Visit}
\item{ AVISITN }{Analysis Visit (N)}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ PARAMN }{Parameter (N)}
\item{ PARCAT1 }{Parameter Category 1}
\item{ AVAL }{Analysis Value}
\item{ AVALC }{Analysis Value (C)}
\item{ BASE }{Baseline Value}
\item{ BASEC }{Baseline Value (C)}
\item{ BASETYPE }{Baseline Type}
\item{ CHG }{Change from Baseline}
\item{ PCHG }{Percent Change from Baseline}
\item{ R2BASE }{Ratio to Baseline}
\item{ R2ANRLO }{Ratio of Analysis Val compared to ANRLO}
\item{ R2ANRHI }{Ratio of Analysis Val compared to ANRHI}
\item{ SHIFT1 }{Shift from Baseline to Analysis Value}
\item{ SHIFT2 }{Shift from Baseline to Overall Grade}
\item{ DTYPE }{Derivation Type}
\item{ ATOXGR }{Analysis Toxicity Grade}
\item{ BTOXGR }{Baseline Toxicity Grade}
\item{ ANRIND }{Analysis Reference Range Indicator}
\item{ BNRIND }{Baseline Reference Range Indicator}
\item{ ANRLO }{Analysis Normal Range Lower Limit}
\item{ ANRHI }{Analysis Normal Range Upper Limit}
\item{ ATOXGRL }{Analysis Toxicity Grade Low}
\item{ ATOXGRH }{Analysis Toxicity Grade High}
\item{ BTOXGRL }{Baseline Toxicity Grade Low}
\item{ BTOXGRH }{Baseline Toxicity Grade High}
\item{ ATOXDSCL }{Analysis Toxicity Description Low}
\item{ ATOXDSCH }{Analysis Toxicity Description High}
\item{ ABLFL }{Baseline Record Flag}
\item{ ANL01FL }{Analysis Flag 01}
\item{ ONTRTFL }{On Treatment Record Flag}
\item{ LVOTFL }{Last Value On Treatment Record Flag}
\item{ LBSEQ }{Sequence Number}
\item{ LBTESTCD }{Lab Test or Examination Short Name}
\item{ LBTEST }{Lab Test or Examination Name}
\item{ LBCAT }{Category for Lab Test}
\item{ LBORRES }{Result or Finding in Original Units}
\item{ LBORRESU }{Original Units}
\item{ LBORNRLO }{Reference Range Lower Limit in Orig Unit}
\item{ LBORNRHI }{Reference Range Upper Limit in Orig Unit}
\item{ LBSTRESC }{Character Result/Finding in Std Format}
\item{ LBSTRESN }{Numeric Result/Finding in Standard Units}
\item{ LBSTRESU }{Standard Units}
\item{ LBSTNRLO }{Reference Range Lower Limit-Std Units}
\item{ LBSTNRHI }{Reference Range Upper Limit-Std Units}
\item{ LBNRIND }{Reference Range Indicator}
\item{ LBBLFL }{Baseline Flag}
\item{ VISITNUM }{Visit Number}
\item{ VISIT }{Visit Name}
\item{ VISITDY }{Planned Study Day of Visit}
\item{ LBDTC }{Date/Time of Specimen Collection}
\item{ LBDY }{Study Day of Specimen Collection}
}
}
\source{
Generated from admiral package (template ad_adlb.R).
}
\usage{
adlb
}
\description{
Laboratory Analysis
}
\details{
Contains a set of 47 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
ALB     \tab Albumin (g/L)                                   \cr
ALKPH   \tab Alkaline Phosphatase (U/L)                      \cr
ALT     \tab Alanine Aminotransferase (U/L)                  \cr
ANISO   \tab Anisocytes                                      \cr
AST     \tab Aspartate Aminotransferase (U/L)                \cr
BASO    \tab Basophils Abs (10^9/L)                          \cr
BASOLE  \tab Basophils/Leukocytes (FRACTION)                 \cr
BILI    \tab Bilirubin (umol/L)                              \cr
BUN     \tab Blood Urea Nitrogen (mmol/L)                    \cr
CA      \tab Calcium (mmol/L)                                \cr
CHOLES  \tab Cholesterol (mmol/L)                            \cr
CK      \tab Creatinine Kinase (U/L)                         \cr
CL      \tab Chloride (mmol/L)                               \cr
COLOR   \tab Color                                           \cr
CREAT   \tab Creatinine (umol/L)                             \cr
EOS     \tab Eosinophils (10^9/L)                            \cr
EOSLE   \tab Eosinophils/Leukocytes (FRACTION)               \cr
GGT     \tab Gamma Glutamyl Transferase (U/L)                \cr
GLUC    \tab Glucose (mmol/L)                                \cr
HBA1C   \tab Hemoglobin A1C (1)                              \cr
HCT     \tab Hematocrit (1)                                  \cr
HGB     \tab Hemoglobin (mmol/L)                             \cr
KETON   \tab Ketones                                         \cr
LYMPH   \tab Lymphocytes Abs (10^9/L)                        \cr
LYMPHLE \tab Lymphocytes/Leukocytes (FRACTION)               \cr
MACROC  \tab Macrocytes                                      \cr
MCH     \tab Ery. Mean Corpuscular Hemoglobin (fmol(Fe))     \cr
MCHC    \tab Ery. Mean Corpuscular HGB Concentration (mmol/L)\cr
MCV     \tab Ery. Mean Corpuscular Volume (f/L)              \cr
MICROC  \tab Microcytes                                      \cr
MONO    \tab Monocytes (10^9/L)                              \cr
MONOLE  \tab Monocytes/Leukocytes (FRACTION)                 \cr
PH      \tab pH                                              \cr
PHOS    \tab Phosphate (mmol/L)                              \cr
PLAT    \tab Platelet (10^9/L)                               \cr
POIKIL  \tab Poikilocytes                                    \cr
POLYCH  \tab Polychromasia                                   \cr
POTAS   \tab Potassium (mmol/L)                              \cr
PROT    \tab Protein (g/L)                                   \cr
RBC     \tab Erythrocytes (TI/L)                             \cr
SODIUM  \tab Sodium (mmol/L)                                 \cr
SPGRAV  \tab Specific Gravity                                \cr
TSH     \tab Thyrotropin (mU/L)                              \cr
URATE   \tab Urate (umol/L)                                  \cr
UROBIL  \tab Urobilinogen                                    \cr
VITB12  \tab Vitamin B12 (pmol/L)                            \cr
WBC     \tab Leukocytes (10^9/L)
}
}
\examples{
data("adlb")
}
\references{
None
}
\keyword{dataset}
\keyword{generic}
