% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb_neuro.R
\docType{data}
\name{adlb_neuro}
\alias{adlb_neuro}
\title{Laboratory Analysis for Neuroscience}
\format{
A data frame with 48 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ DOMAIN }{Domain Abbreviation}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ AVISIT }{Analysis Visit}
\item{ AVISITN }{Analysis Visit (N)}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ PARAMN }{Parameter (N)}
\item{ AVAL }{Analysis Value}
\item{ AVALC }{Analysis Value (C)}
\item{ ANRLO }{Analysis Normal Range Lower Limit}
\item{ ANRHI }{Analysis Normal Range Upper Limit}
\item{ BASE }{Baseline Value}
\item{ BASEC }{Baseline Value (C)}
\item{ BASETYPE }{Baseline Type}
\item{ CHG }{Change from Baseline}
\item{ PCHG }{Percent Change from Baseline}
\item{ ABLFL }{Baseline Record Flag}
\item{ ANL01FL }{Analysis Flag 01}
\item{ ANL02FL }{Analysis Flag 02}
\item{ ONTRTFL }{On Treatment Record Flag}
\item{ ASEQ }{Analysis Sequence Number}
\item{ LBSEQ }{Sequence Number}
\item{ LBTESTCD }{Lab Test or Examination Short Name}
\item{ LBTEST }{Lab Test or Examination Name}
\item{ LBCAT }{Category for Lab Test}
\item{ LBORRES }{Result or Finding in Original Units}
\item{ LBORRESU }{Original Units}
\item{ LBORNRLO }{Reference Range Lower Limit in Orig Unit}
\item{ LBORNRHI }{Reference Range Upper Limit in Orig Unit}
\item{ LBSTRESC }{Character Result/Finding in Std Format}
\item{ LBSTRESN }{Numeric Result/Finding in Standard Units}
\item{ LBSTRESU }{Standard Units}
\item{ LBSTNRLO }{Reference Range Lower Limit-Std Units}
\item{ LBSTNRHI }{Reference Range Upper Limit-Std Units}
\item{ LBNRIND }{Reference Range Indicator}
\item{ LBBLFL }{Baseline Flag}
\item{ VISITNUM }{Visit Number}
\item{ VISIT }{Visit Name}
\item{ VISITDY }{Planned Study Day of Visit}
\item{ LBDTC }{Date/Time of Specimen Collection}
\item{ LBDY }{Study Day of Specimen Collection}
}
}
\source{
Generated from admiralneuro package (template ad_adlb.R).
}
\usage{
adlb_neuro
}
\description{
Laboratory Analysis for Neuroscience
}
\details{
Contains a set of 6 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
AMYLB42  \tab Lumipulse G Beta-Amyloid 1-42-N Plasma (pg/mL)                \cr
ASYNASAA \tab Alpha Synuclein Seed Amplification Assay (CSF)                \cr
LAMYLB42 \tab Log-Transformed Lumipulse G Beta-Amyloid 1-42-N Plasma (pg/mL)\cr
PTAB42R  \tab Lumipulse G pTau 217/Beta-Amyloid 1-42 Plasma Ratio           \cr
PTAU217  \tab Lumipulse G pTau 217 Plasma (pg/mL)                           \cr
TAU181P  \tab Elecsys Tau Protein Phosphorylated 181
}
}
\examples{
data("adlb_neuro")
}
\references{
None
}
\keyword{dataset}
\keyword{neurology}
