% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlbhy.R
\docType{data}
\name{adlbhy}
\alias{adlbhy}
\title{Analysis of Lab Hy's Law}
\format{
A data frame with 14 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ PARAMCD }{Parameter Code}
\item{ PARAM }{Parameter}
\item{ LBSEQ }{Sequence Number}
\item{ ADT }{Analysis Date}
\item{ AVISIT }{Analysis Visit}
\item{ ADY }{Analysis Relative Day}
\item{ AVAL }{Analysis Value}
\item{ ANRHI }{Analysis Normal Range Upper Limit}
\item{ CRIT1 }{Analysis Criterion 1}
\item{ CRIT1FL }{Criterion 1 Evaluation Result Flag}
\item{ AVALC }{Analysis Value (C)}
}
}
\source{
Generated from admiral package (template ad_adlbhy.R).
}
\usage{
adlbhy
}
\description{
Analysis of Lab Hy's Law
}
\details{
Contains a set of 4 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
ALT    \tab Alanine Aminotransferase (U/L)    \cr
AST    \tab Aspartate Aminotransferase (U/L)  \cr
BILI   \tab Bilirubin (umol/L)                \cr
HYSLAW \tab ALT/AST >= 3xULN and BILI >= 2xULN
}
}
\examples{
data("adlbhy")
}
\references{
None
}
\keyword{datasets}
