% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admh.R
\docType{data}
\name{admh}
\alias{admh}
\title{Medical History Analysis}
\format{
A data frame with 114 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRTP }{Planned Treatment}
\item{ TRTA }{Actual Treatment}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Treatment End Datetime Imput Flag}
\item{ APHASE }{Phase}
\item{ APHASEN }{Description of Phase N}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHDTF }{Date of Death Imputation Flag}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ MHSEQ }{Sequence Number}
\item{ MHTERM }{Reported Term for the Medical History}
\item{ MHTERMN }{Medical History Term (N)}
\item{ MHDECOD }{Dictionary-Derived Term}
\item{ MHBODSYS }{Body System or Organ Class}
\item{ MHLLT }{Lowest Level Term}
\item{ MHHLT }{High Level Term}
\item{ MHHLGT }{High Level Group Term}
\item{ MHCAT }{Category for Medical History}
\item{ MHSTDTC }{Start Date/Time of Medical History Event}
\item{ ASTDT }{Analysis Start Date}
\item{ MHENDTC }{End Date/Time of Medical History Event}
\item{ AENDT }{Analysis End Date}
\item{ ASTDY }{Analysis Start Relative Day}
\item{ AENDY }{Analysis End Relative Day}
\item{ MHOCCUR }{Medical History Occurrence}
\item{ MHPRESP }{Medical History Event Pre-Specified}
\item{ ANL01FL }{Analysis Flag 01}
\item{ AOCCFL }{1st Occurrence within Subject Flag}
\item{ AOCCPFL }{1st Occurrence of Preferred Term Flag}
\item{ AOCCSFL }{1st Occurrence of SOC Flag}
\item{ MHSPID }{Sponsor-Defined Identifier}
\item{ MHSEV }{Severity/Intensity}
\item{ VISITNUM }{Visit Number}
\item{ VISIT }{Visit Name}
\item{ VISITDY }{Planned Study Day of Visit}
\item{ MHDTC }{Date/Time of History Collection}
\item{ MHDY }{Study Day of History Collection}
\item{ MHSTRTPT }{Start Relative to Reference Time Point}
\item{ MHENRTPT }{End Relative to Reference Time Point}
\item{ MHSTTPT }{Start Reference Time Point}
\item{ MHENTPT }{End Reference Time Point}
\item{ MHENRF }{End Relative to Reference Period}
\item{ MHSTAT }{Completion Status}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ SMQ02NAM }{SMQ 02 Name}
\item{ SMQ02CD }{SMQ 02 Code}
\item{ SMQ02SC }{SMQ 02 Scope}
\item{ SMQ02SCN }{SMQ 02 Scope (N)}
\item{ SMQ03NAM }{SMQ 03 Name}
\item{ SMQ03CD }{SMQ 03 Code}
\item{ SMQ03SC }{SMQ 03 Scope}
\item{ SMQ03SCN }{SMQ 03 Scope (N)}
\item{ SMQ05NAM }{SMQ 05 Name}
\item{ SMQ05CD }{SMQ 05 Code}
\item{ SMQ05SC }{SMQ 05 Scope}
\item{ SMQ05SCN }{SMQ 05 Scope (N)}
\item{ CQ01NAM }{Customized Query 01 Name}
\item{ CQ04NAM }{Customized Query 04 Name}
\item{ CQ04CD }{Customized Query 04 Code}
\item{ AHIST }{Response of Med Hx (past or current)}
\item{ AOCPFL }{1st Occur w/in Trt Prd FL}
\item{ AOCPSFL }{1st Occur of SOC w/in Trt Prd FL}
\item{ AOCPPFL }{1st Occur of PT w/in Trt Prd FL}
}
}
\source{
Generated from admiral package (template ad_admh.R).
}
\usage{
admh
}
\description{
Medical History Analysis
}
\examples{
data("admh")
}
\references{
None
}
\keyword{datasets}
