% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpp.R
\docType{data}
\name{adpp}
\alias{adpp}
\title{Pharmacokinetic Parameters}
\format{
A data frame with 79 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRTP }{Planned Treatment}
\item{ TRTA }{Actual Treatment}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Time of Last Exposure Imput. Flag}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHDTF }{Date of Death Imputation Flag}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ AVISIT }{Analysis Visit}
\item{ AVISITN }{Analysis Visit (N)}
\item{ PARAMCD }{Parameter Code}
\item{ AVAL }{Numeric Result/Finding in Standard Units}
\item{ AVALCAT1 }{Analysis Value Category 1}
\item{ AVALCA1N }{Analysis Value Category 1 (N)}
\item{ SRCDOM }{Domain Abbreviation}
\item{ SRCVAR }{Source Variable}
\item{ SRCSEQ }{Sequence Number}
\item{ PPTESTCD }{Parameter Short Name}
\item{ PPTEST }{Parameter Name}
\item{ PPCAT }{Parameter Category}
\item{ PPORRES }{Result or Finding in Original Units}
\item{ PPORRESU }{Original Units}
\item{ PPSTRESU }{Standard Units}
\item{ PPSPEC }{Specimen Material Type}
\item{ PPRFDTC }{Date/Time of Reference Point}
\item{ VISIT }{Visit Name}
\item{ VISITNUM }{Visit Number}
\item{ PARCAT1 }{Parameter Category}
\item{ AVALU }{Standard Units}
}
}
\source{
Generated from admiral package (template ad_adpp.R).
}
\usage{
adpp
}
\description{
Pharmacokinetic Parameters
}
\examples{
data("adpp")
}
\references{
None
}
\keyword{dataset}
\keyword{generic}
