% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adppk.R
\docType{data}
\name{adppk}
\alias{adppk}
\title{Dataset adppk}
\format{
A data frame with 59 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ EVID }{EVID}
\item{ NFRLT }{NFRLT}
\item{ AFRLT }{AFRLT}
\item{ APRLT }{APRLT}
\item{ NPRLT }{NPRLT}
\item{ DOSEA }{DOSEA}
\item{ DOSEP }{DOSEP}
\item{ ALLOQ }{ALLOQ}
\item{ CMT }{CMT}
\item{ BLQFL }{BLQFL}
\item{ BLQFN }{BLQFN}
\item{ AMT }{AMT}
\item{ DV }{DV}
\item{ AVAL }{AVAL}
\item{ DVL }{DVL}
\item{ MDV }{MDV}
\item{ AVALU }{AVALU}
\item{ UDTC }{UDTC}
\item{ II }{II}
\item{ SS }{SS}
\item{ ASEQ }{ASEQ}
\item{ PROJID }{PROJID}
\item{ PROJIDN }{PROJIDN}
\item{ STUDYIDN }{STUDYIDN}
\item{ SITEID }{SITEID}
\item{ SITEIDN }{SITEIDN}
\item{ USUBJIDN }{USUBJIDN}
\item{ SUBJID }{SUBJID}
\item{ SUBJIDN }{SUBJIDN}
\item{ AGE }{AGE}
\item{ SEX }{SEX}
\item{ SEXN }{SEXN}
\item{ COHORT }{COHORT}
\item{ COHORTC }{COHORTC}
\item{ ROUTE }{ROUTE}
\item{ ROUTEN }{ROUTEN}
\item{ RACE }{RACE}
\item{ RACEN }{RACEN}
\item{ ETHNIC }{ETHNIC}
\item{ ETHNICN }{ETHNICN}
\item{ FORM }{FORM}
\item{ FORMN }{FORMN}
\item{ COUNTRY }{COUNTRY}
\item{ COUNTRYN }{COUNTRYN}
\item{ REGION1 }{REGION1}
\item{ REGION1N }{REGION1N}
\item{ HTBL }{HTBL}
\item{ WTBL }{WTBL}
\item{ ALTBL }{ALTBL}
\item{ ASTBL }{ASTBL}
\item{ TBILBL }{TBILBL}
\item{ CREATBL }{CREATBL}
\item{ BMIBL }{BMIBL}
\item{ BSABL }{BSABL}
\item{ CRCLBL }{CRCLBL}
\item{ EGFRBL }{EGFRBL}
\item{ RECSEQ }{RECSEQ}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adppk
}
\description{
adppk dataset
}
\examples{
load(adppk)
}
\references{
None
}
\keyword{datasets}
