% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrs_onco.R
\docType{data}
\name{adrs_onco}
\alias{adrs_onco}
\title{Dataset adrs_onco}
\format{
A data frame with 75 columns:
\describe{
\item{ DOMAIN }{DOMAIN}
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ VISITNUM }{VISITNUM}
\item{ VISIT }{VISIT}
\item{ RSTESTCD }{RSTESTCD}
\item{ RSTEST }{RSTEST}
\item{ RSORRES }{RSORRES}
\item{ RSSTRESC }{RSSTRESC}
\item{ RSEVAL }{RSEVAL}
\item{ RSEVALID }{RSEVALID}
\item{ RSACPTFL }{RSACPTFL}
\item{ RSDTC }{RSDTC}
\item{ RSSEQ }{RSSEQ}
\item{ RANDDT }{RANDDT}
\item{ PARAMCD }{PARAMCD}
\item{ PARAM }{PARAM}
\item{ PARCAT1 }{PARCAT1}
\item{ PARCAT2 }{PARCAT2}
\item{ PARCAT3 }{PARCAT3}
\item{ ADT }{ADT}
\item{ ADTF }{ADTF}
\item{ AVISIT }{AVISIT}
\item{ AVALC }{AVALC}
\item{ AVAL }{AVAL}
\item{ ANL01FL }{ANL01FL}
\item{ ANL02FL }{ANL02FL}
\item{ ASEQ }{ASEQ}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P }{TRT01P}
\item{ TRT01A }{TRT01A}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from create_adams_data.R script.
}
\usage{
adrs_onco
}
\description{
adrs_onco dataset
}
\examples{
data("adrs_onco")
}
\references{
None
}
\keyword{datasets}
