% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Dataset adsl}
\format{
A data frame with 51 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P }{TRT01P}
\item{ TRT01A }{TRT01A}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT }{DTHDT}
\item{ DTHDTF }{DTHDTF}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ AGEGR1 }{AGEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adsl
}
\description{
adsl dataset
}
\examples{
data("adsl")
}
\references{
None
}
\keyword{datasets}
