% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_vaccine.R
\docType{data}
\name{adsl_vaccine}
\alias{adsl_vaccine}
\title{Subject Level Analysis for Vaccine}
\format{
A data frame with 48 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country/Region}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ REGION1 }{Geographic Region 1}
\item{ BRTHDTC }{Date/Time of Birth}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ PPROTFL }{Per-Protocol Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRT02P }{Planned Treatment for Period 02}
\item{ TRT02A }{Actual Treatment for Period 02}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ AP01SDT }{Period 01 Start Date}
\item{ AP01EDT }{Period 01 End Date}
\item{ AP02SDT }{Period 02 Start Date}
\item{ AP02EDT }{Period 02 End Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ INVID }{Investigator Identifier}
\item{ INVNAM }{Investigator Name}
\item{ VAX01DT }{Vaccination Date 01}
\item{ VAX02DT }{Vaccination Date 02}
\item{ ARMNRS }{Reason Arm and/or Actual Arm is Null}
\item{ ACTARMUD }{Description of Unplanned Actual Arm}
}
}
\source{
Generated from admiralvaccine package (template ad_adsl.R).
}
\usage{
adsl_vaccine
}
\description{
Subject Level Analysis for Vaccine
}
\examples{
data("adsl_vaccine")
}
\references{
None
}
\keyword{dataset}
\keyword{vaccine}
