% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtpet_neuro.R
\docType{data}
\name{adtpet_neuro}
\alias{adtpet_neuro}
\title{Tau PET Scan Analysis Dataset}
\format{
A data frame with 46 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ DOMAIN }{Domain Abbreviation}
\item{ ASEQ }{Analysis Sequence Number}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ AVISIT }{Analysis Visit}
\item{ AVISITN }{Analysis Visit (N)}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ AVAL }{Analysis Value}
\item{ AVALC }{Analysis Value (C)}
\item{ BASE }{Baseline Value}
\item{ BASEC }{Baseline Value (C)}
\item{ BASETYPE }{Baseline Type}
\item{ CHG }{Change from Baseline}
\item{ PCHG }{Percent Change from Baseline}
\item{ ABLFL }{Baseline Record Flag}
\item{ ANL01FL }{Analysis Flag 01}
\item{ ANL02FL }{Analysis Flag 02}
\item{ ONTRTFL }{On Treatment Record Flag}
\item{ NVSEQ }{Sequence Number}
\item{ NVLNKID }{Link ID}
\item{ NVTESTCD }{Short Name of Nervous System Test}
\item{ NVTEST }{Name of Nervous System Test}
\item{ NVCAT }{Category for Nervous System Test}
\item{ NVLOC }{Location Used for the Measurement}
\item{ NVNAM }{Vendor Name}
\item{ NVORRES }{Result or Finding in Original Units}
\item{ NVORRESU }{Original Units}
\item{ NVSTRESC }{Character Result/Finding in Std Format}
\item{ NVSTRESN }{Numeric Result/Finding in Standard Units}
\item{ NVSTRESU }{Standard Units}
\item{ NVMETHOD }{Method of Test or Examination}
\item{ NVLOBXFL }{Last Observation Before Exposure Flag}
\item{ REFREG }{Reference Region}
\item{ AGTRT }{Reported Agent Name}
\item{ AGCAT }{Category for Agent}
\item{ VISITNUM }{Visit Number}
\item{ VISIT }{Visit Name}
\item{ NVDTC }{Date/Time of Collection}
\item{ NVDY }{Study Day of Visit/Collection/Exam}
}
}
\source{
Generated from admiralneuro package (template ad_adtpet.R).
}
\usage{
adtpet_neuro
}
\description{
Tau PET Scan Analysis Dataset
}
\details{
Contains a set of 2 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
SUVRAFTP \tab AVID FTP Standard Uptake Ratio Neocortical Composite Inferior Cerebellar Gray Matter    \cr
SUVRBFTP \tab BERKELEY FTP Standard Uptake Ratio Neocortical Composite Inferior Cerebellar Gray Matter
}
}
\examples{
data("adtpet_neuro")
}
\references{
None
}
\keyword{dataset}
\keyword{neurology}
