% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtr_onco.R
\docType{data}
\name{adtr_onco}
\alias{adtr_onco}
\title{Dataset adtr_onco}
\format{
A data frame with 100 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ DOMAIN }{DOMAIN}
\item{ USUBJID }{USUBJID}
\item{ TRSEQ }{TRSEQ}
\item{ TRGRPID }{TRGRPID}
\item{ TRLNKGRP }{TRLNKGRP}
\item{ TRLNKID }{TRLNKID}
\item{ TRTESTCD }{TRTESTCD}
\item{ TRTEST }{TRTEST}
\item{ TRORRES }{TRORRES}
\item{ TRORRESU }{TRORRESU}
\item{ TRSTRESC }{TRSTRESC}
\item{ TRSTRESN }{TRSTRESN}
\item{ TRSTRESU }{TRSTRESU}
\item{ TRSTAT }{TRSTAT}
\item{ TRREASND }{TRREASND}
\item{ TRMETHOD }{TRMETHOD}
\item{ TREVAL }{TREVAL}
\item{ TREVALID }{TREVALID}
\item{ TRACPTFL }{TRACPTFL}
\item{ VISITNUM }{VISITNUM}
\item{ VISIT }{VISIT}
\item{ TRDTC }{TRDTC}
\item{ TRDY }{TRDY}
\item{ RANDDT }{RANDDT}
\item{ TULOC }{TULOC}
\item{ TULOCGR1 }{TULOCGR1}
\item{ LSEXP }{LSEXP}
\item{ LSASS }{LSASS}
\item{ ADT }{ADT}
\item{ ADTF }{ADTF}
\item{ ADY }{ADY}
\item{ AVISIT }{AVISIT}
\item{ AVISITN }{AVISITN}
\item{ PARAMCD }{PARAMCD}
\item{ PARAM }{PARAM}
\item{ PARCAT1 }{PARCAT1}
\item{ PARCAT2 }{PARCAT2}
\item{ PARCAT3 }{PARCAT3}
\item{ AVAL }{AVAL}
\item{ ANL01FL }{ANL01FL}
\item{ ABLFL }{ABLFL}
\item{ BASE }{BASE}
\item{ NADIR }{NADIR}
\item{ CHG }{CHG}
\item{ PCHG }{PCHG}
\item{ CHGNAD }{CHGNAD}
\item{ PCHGNAD }{PCHGNAD}
\item{ PDFL }{PDFL}
\item{ ANL02FL }{ANL02FL}
\item{ ANL03FL }{ANL03FL}
\item{ ANL04FL }{ANL04FL}
\item{ ASEQ }{ASEQ}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P }{TRT01P}
\item{ TRT01A }{TRT01A}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiralonco.
}
\usage{
adtr_onco
}
\description{
adtr_onco dataset
}
\examples{
data("adtr_onco")
}
\references{
None
}
\keyword{datasets}
