% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtte_onco.R
\docType{data}
\name{adtte_onco}
\alias{adtte_onco}
\title{Dataset adtte_onco}
\format{
A data frame with 20 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ EVNTDESC }{EVNTDESC}
\item{ SRCDOM }{SRCDOM}
\item{ SRCVAR }{SRCVAR}
\item{ SRCSEQ }{SRCSEQ}
\item{ CNSR }{CNSR}
\item{ ADT }{ADT}
\item{ CNSDTDSC }{CNSDTDSC}
\item{ STARTDT }{STARTDT}
\item{ PARAMCD }{PARAMCD}
\item{ PARAM }{PARAM}
\item{ AVAL }{AVAL}
\item{ ASEQ }{ASEQ}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ AGE }{AGE}
\item{ SEX }{SEX}
}
}
\source{
Generated from create_adams_data.R script.
}
\usage{
adtte_onco
}
\description{
adtte_onco dataset
}
\examples{
data("adtte_onco")
}
\references{
None
}
\keyword{datasets}
