% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advfq_admiralophtha.R
\docType{data}
\name{advfq_admiralophtha}
\alias{advfq_admiralophtha}
\title{Dataset advfq_admiralophtha}
\format{
A data frame with 85 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ DOMAIN }{DOMAIN}
\item{ USUBJID }{USUBJID}
\item{ QSSEQ }{QSSEQ}
\item{ QSTESTCD }{QSTESTCD}
\item{ QSTEST }{QSTEST}
\item{ QSCAT }{QSCAT}
\item{ QSSCAT }{QSSCAT}
\item{ QSORRES }{QSORRES}
\item{ QSORRESU }{QSORRESU}
\item{ QSSTRESC }{QSSTRESC}
\item{ QSSTRESN }{QSSTRESN}
\item{ QSSTRESU }{QSSTRESU}
\item{ QSBLFL }{QSBLFL}
\item{ QSDRVFL }{QSDRVFL}
\item{ VISITNUM }{VISITNUM}
\item{ VISIT }{VISIT}
\item{ VISITDY }{VISITDY}
\item{ QSDTC }{QSDTC}
\item{ QSDY }{QSDY}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRT01A }{TRT01A}
\item{ TRT01P }{TRT01P}
\item{ ADT }{ADT}
\item{ ADY }{ADY}
\item{ PARAMCD }{PARAMCD}
\item{ AVAL }{AVAL}
\item{ AVALC }{AVALC}
\item{ AVISIT }{AVISIT}
\item{ AVISITN }{AVISITN}
\item{ ONTRTFL }{ONTRTFL}
\item{ ABLFL }{ABLFL}
\item{ BASE }{BASE}
\item{ CHG }{CHG}
\item{ PCHG }{PCHG}
\item{ ANL01FL }{ANL01FL}
\item{ ASEQ }{ASEQ}
\item{ PARAM }{PARAM}
\item{ PARCAT1 }{PARCAT1}
\item{ PARCAT2 }{PARCAT2}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from create_adams_data.R script.
}
\usage{
advfq_admiralophtha
}
\description{
advfq_admiralophtha dataset
}
\examples{
data("advfq_admiralophtha")
}
\references{
None
}
\keyword{datasets}
