% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advfq_ophtha.R
\docType{data}
\name{advfq_ophtha}
\alias{advfq_ophtha}
\title{Dataset advfq_ophtha}
\format{
A data frame with 85 columns:
\describe{
\item{ STUDYID }{undocumented field}
\item{ DOMAIN }{undocumented field}
\item{ USUBJID }{undocumented field}
\item{ QSSEQ }{undocumented field}
\item{ QSTESTCD }{undocumented field}
\item{ QSTEST }{undocumented field}
\item{ QSCAT }{undocumented field}
\item{ QSSCAT }{undocumented field}
\item{ QSORRES }{undocumented field}
\item{ QSORRESU }{undocumented field}
\item{ QSSTRESC }{undocumented field}
\item{ QSSTRESN }{undocumented field}
\item{ QSSTRESU }{undocumented field}
\item{ QSBLFL }{undocumented field}
\item{ QSDRVFL }{undocumented field}
\item{ VISITNUM }{undocumented field}
\item{ VISIT }{undocumented field}
\item{ VISITDY }{undocumented field}
\item{ QSDTC }{undocumented field}
\item{ QSDY }{undocumented field}
\item{ TRTSDT }{undocumented field}
\item{ TRTEDT }{undocumented field}
\item{ TRT01A }{undocumented field}
\item{ TRT01P }{undocumented field}
\item{ ADT }{undocumented field}
\item{ ADY }{undocumented field}
\item{ PARAMCD }{undocumented field}
\item{ AVAL }{undocumented field}
\item{ AVALC }{undocumented field}
\item{ AVISIT }{undocumented field}
\item{ AVISITN }{undocumented field}
\item{ ONTRTFL }{undocumented field}
\item{ ABLFL }{undocumented field}
\item{ BASE }{undocumented field}
\item{ CHG }{undocumented field}
\item{ PCHG }{undocumented field}
\item{ ANL01FL }{undocumented field}
\item{ ASEQ }{undocumented field}
\item{ PARAM }{undocumented field}
\item{ PARCAT1 }{undocumented field}
\item{ PARCAT2 }{undocumented field}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ SITEID }{Study Site Identifier}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ ETHNIC }{Ethnicity}
\item{ ARMCD }{Planned Arm Code}
\item{ ARM }{Description of Planned Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ COUNTRY }{Country}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ TRTSDTM }{undocumented field}
\item{ TRTSTMF }{undocumented field}
\item{ TRTEDTM }{undocumented field}
\item{ TRTETMF }{undocumented field}
\item{ TRTDURD }{undocumented field}
\item{ SCRFDT }{undocumented field}
\item{ EOSDT }{undocumented field}
\item{ EOSSTT }{undocumented field}
\item{ FRVDT }{undocumented field}
\item{ RANDDT }{undocumented field}
\item{ DTHDT }{undocumented field}
\item{ DTHADY }{undocumented field}
\item{ LDDTHELD }{undocumented field}
\item{ LSTALVDT }{undocumented field}
\item{ AGEGR1 }{undocumented field}
\item{ SAFFL }{undocumented field}
\item{ RACEGR1 }{undocumented field}
\item{ REGION1 }{undocumented field}
\item{ LDDTHGR1 }{undocumented field}
\item{ DTH30FL }{undocumented field}
\item{ DTHA30FL }{undocumented field}
\item{ DTHB30FL }{undocumented field}
}
}
\source{
Generated from admiralophtha package (template ad_advfq.R).
}
\usage{
advfq_ophtha
}
\description{
advfq_ophtha dataset
}
\examples{
data("advfq_ophtha")
}
\references{
None
}
\keyword{datasets}
