% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcoeq_metabolic.R
\docType{data}
\name{adcoeq_metabolic}
\alias{adcoeq_metabolic}
\title{Questionnaires Analysis for Metabolic}
\format{
A data frame with 85 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ ASEQ }{Analysis Sequence Number}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Time of Last Exposure Imput. Flag}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHDTF }{Date of Death Imputation Flag}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ ADT }{Analysis Date}
\item{ ADY }{Analysis Relative Day}
\item{ AVISIT }{Analysis Visit}
\item{ AVISITN }{Analysis Visit (N)}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ PARAMN }{Parameter (N)}
\item{ PARCAT1 }{Parameter Category 1}
\item{ AVAL }{Analysis Value}
\item{ AVALC }{Analysis Value (C)}
\item{ BASE }{Baseline Value}
\item{ CHG }{Change from Baseline}
\item{ PCHG }{Percent Change from Baseline}
\item{ ABLFL }{Baseline Record Flag}
\item{ VISIT }{Visit Name}
\item{ VISITNUM }{Visit Number}
\item{ VISITDY }{Planned Study Day of Visit}
\item{ QSBLFL }{Baseline Flag}
\item{ QSDTC }{Date/Time of Finding}
\item{ QSDY }{Study Day of Finding}
\item{ QSCAT }{Category for Questionnaire}
\item{ QSTEST }{Questionnaire Test Name}
\item{ QSTESTCD }{Questionnaire Test Short Name}
\item{ QSORRES }{Result or Finding in Original Units}
\item{ QSORRESU }{Original Units}
\item{ QSSTRESC }{Character Result/Finding in Std Format}
\item{ QSSTRESN }{Numeric Result/Finding in Standard Units}
\item{ QSSTRESU }{Standard Units}
\item{ QSSEQ }{Sequence Number}
}
}
\source{
Generated from admiralmetabolic package (template ad_adcoeq.R).
}
\usage{
adcoeq_metabolic
}
\description{
Questionnaires Analysis for Metabolic
}
\details{
Contains a set of 25 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
COEQ01   \tab How hungry have you felt?                                                   \cr
COEQ02   \tab How full have you felt?                                                     \cr
COEQ03   \tab How strong was your desire to eat sweet foods?                              \cr
COEQ04   \tab How strong was your desire to eat savoury foods?                            \cr
COEQ05   \tab How happy have you felt?                                                    \cr
COEQ06   \tab How anxious have you felt?                                                  \cr
COEQ07   \tab How alert have you felt?                                                    \cr
COEQ08   \tab How contented have you felt?                                                \cr
COEQ09   \tab During the last 7 days how often have you had food cravings?                \cr
COEQ10   \tab How strong have any food cravings been?                                     \cr
COEQ11   \tab How difficult has it been to resist any food cravings?                      \cr
COEQ12   \tab How often have you eaten in response to food cravings?                      \cr
COEQ13   \tab Chocolate or chocolate flavoured foods                                      \cr
COEQ14   \tab Other sweet foods (cakes, pastries, biscuits, etc)                          \cr
COEQ15   \tab Fruit or fruit juice                                                        \cr
COEQ16   \tab Dairy foods (cheese, yoghurts, milk, etc)                                   \cr
COEQ17   \tab Starchy foods (bread, rice, pasta, etc)                                     \cr
COEQ18   \tab Savoury foods (french fries, crisps, burgers, pizza, etc)                   \cr
COEQ19   \tab Generally, how difficult has it been to control your eating?                \cr
COEQ20   \tab Which one food makes it most difficult for you to control eating?           \cr
COEQ21   \tab How difficult has it been to resist eating this food during the last 7 days?\cr
COEQCRCO \tab COEQ - Craving Control                                                      \cr
COEQCRSA \tab COEQ - Craving for Savoury                                                  \cr
COEQCRSW \tab COEQ - Craving for Sweet                                                    \cr
COEQPOMO \tab COEQ - Positive Mood
}
}
\examples{
data("adcoeq_metabolic")
}
\references{
None
}
\keyword{datasets}
