% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeg.R
\docType{data}
\name{adeg}
\alias{adeg}
\title{Dataset adeg}
\format{
A data frame with 105 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ DOMAIN }{DOMAIN}
\item{ USUBJID }{USUBJID}
\item{ EGSEQ }{EGSEQ}
\item{ EGTESTCD }{EGTESTCD}
\item{ EGTEST }{EGTEST}
\item{ EGORRES }{EGORRES}
\item{ EGORRESU }{EGORRESU}
\item{ EGSTRESC }{EGSTRESC}
\item{ EGSTRESN }{EGSTRESN}
\item{ EGSTRESU }{EGSTRESU}
\item{ EGSTAT }{EGSTAT}
\item{ EGLOC }{EGLOC}
\item{ EGBLFL }{EGBLFL}
\item{ VISITNUM }{VISITNUM}
\item{ VISIT }{VISIT}
\item{ VISITDY }{VISITDY}
\item{ EGDTC }{EGDTC}
\item{ EGDY }{EGDY}
\item{ EGTPT }{EGTPT}
\item{ EGTPTNUM }{EGTPTNUM}
\item{ EGELTM }{EGELTM}
\item{ EGTPTREF }{EGTPTREF}
\item{ temp }{temp}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRT01A }{TRT01A}
\item{ TRT01P }{TRT01P}
\item{ ADTM }{ADTM}
\item{ ATMF }{ATMF}
\item{ ADY }{ADY}
\item{ PARAMCD }{PARAMCD}
\item{ AVAL }{AVAL}
\item{ AVALC }{AVALC}
\item{ ADT }{ADT}
\item{ ATPTN }{ATPTN}
\item{ ATPT }{ATPT}
\item{ AVISIT }{AVISIT}
\item{ AVISITN }{AVISITN}
\item{ DTYPE }{DTYPE}
\item{ ONTRTFL }{ONTRTFL}
\item{ ANRLO }{ANRLO}
\item{ ANRHI }{ANRHI}
\item{ ANRIND }{ANRIND}
\item{ BASETYPE }{BASETYPE}
\item{ ABLFL }{ABLFL}
\item{ BASE }{BASE}
\item{ BASEC }{BASEC}
\item{ BNRIND }{BNRIND}
\item{ CHG }{CHG}
\item{ PCHG }{PCHG}
\item{ ANL01FL }{ANL01FL}
\item{ TRTP }{TRTP}
\item{ TRTA }{TRTA}
\item{ ASEQ }{ASEQ}
\item{ AVALCA1N }{AVALCA1N}
\item{ AVALCAT1 }{AVALCAT1}
\item{ CHGCAT1N }{CHGCAT1N}
\item{ CHGCAT1 }{CHGCAT1}
\item{ PARAM }{PARAM}
\item{ PARAMN }{PARAMN}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adeg
}
\description{
adeg dataset
}
\examples{
load("adeg.rda")
}
\references{
None
}
\keyword{datasets}
