% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adis_vaccine.R
\docType{data}
\name{adis_vaccine}
\alias{adis_vaccine}
\title{Dataset adis_vaccine}
\format{
A data frame with 103 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ DOMAIN }{DOMAIN}
\item{ USUBJID }{USUBJID}
\item{ ISSEQ }{ISSEQ}
\item{ ISTESTCD }{ISTESTCD}
\item{ ISTEST }{ISTEST}
\item{ ISCAT }{ISCAT}
\item{ ISORRES }{ISORRES}
\item{ ISORRESU }{ISORRESU}
\item{ ISSTRESC }{ISSTRESC}
\item{ ISSTRESN }{ISSTRESN}
\item{ ISSTRESU }{ISSTRESU}
\item{ ISSTAT }{ISSTAT}
\item{ ISREASND }{ISREASND}
\item{ ISNAM }{ISNAM}
\item{ ISSPEC }{ISSPEC}
\item{ ISMETHOD }{ISMETHOD}
\item{ ISBLFL }{ISBLFL}
\item{ ISLLOQ }{ISLLOQ}
\item{ VISITNUM }{VISITNUM}
\item{ EPOCH }{EPOCH}
\item{ ISDTC }{ISDTC}
\item{ ISDY }{ISDY}
\item{ ISULOQ }{ISULOQ}
\item{ LOD }{LOD}
\item{ AVISITN }{AVISITN}
\item{ AVISIT }{AVISIT}
\item{ ATPTN }{ATPTN}
\item{ ATPT }{ATPT}
\item{ ATPTREF }{ATPTREF}
\item{ ADT }{ADT}
\item{ RFSTDTC }{RFSTDTC}
\item{ PPROTFL }{PPROTFL}
\item{ ADY }{ADY}
\item{ PARAMCD }{PARAMCD}
\item{ PARAM }{PARAM}
\item{ PARAMN }{PARAMN}
\item{ PARCAT1 }{PARCAT1}
\item{ CUTOFF02 }{CUTOFF02}
\item{ CUTOFF03 }{CUTOFF03}
\item{ AVAL }{AVAL}
\item{ AVALU }{AVALU}
\item{ SERCAT1 }{SERCAT1}
\item{ SERCAT1N }{SERCAT1N}
\item{ DTYPE }{DTYPE}
\item{ BASETYPE }{BASETYPE}
\item{ BASE }{BASE}
\item{ ABLFL }{ABLFL}
\item{ BASECAT1 }{BASECAT1}
\item{ CHG }{CHG}
\item{ R2BASE }{R2BASE}
\item{ CRIT1FL }{CRIT1FL}
\item{ CRIT1FN }{CRIT1FN}
\item{ CRIT1 }{CRIT1}
\item{ APERIOD }{APERIOD}
\item{ APERSDT }{APERSDT}
\item{ APEREDT }{APEREDT}
\item{ TRTA }{TRTA}
\item{ TRTP }{TRTP}
\item{ PPSRFL }{PPSRFL}
\item{ SUBJID }{SUBJID}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ INVID }{INVID}
\item{ INVNAM }{INVNAM}
\item{ BRTHDTC }{BRTHDTC}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P }{TRT01P}
\item{ TRT02P }{TRT02P}
\item{ TRT01A }{TRT01A}
\item{ TRT02A }{TRT02A}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRTDURD }{TRTDURD}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ AGEGR1 }{AGEGR1}
\item{ REGION1 }{REGION1}
\item{ VAX01DT }{VAX01DT}
\item{ VAX02DT }{VAX02DT}
\item{ AP01SDT }{AP01SDT}
\item{ AP01EDT }{AP01EDT}
\item{ AP02SDT }{AP02SDT}
\item{ AP02EDT }{AP02EDT}
}
}
\source{
Generated from pharmaverse/admiralvaccine.
}
\usage{
adis_vaccine
}
\description{
adis_vaccine dataset
}
\examples{
data("adis_vaccine")
}
\references{
None
}
\keyword{datasets}
