% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adis_vaccine.R
\docType{data}
\name{adis_vaccine}
\alias{adis_vaccine}
\title{Immunogenicity Specimen Assessments}
\format{
A data frame with 104 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ DOMAIN }{Domain Abbreviation}
\item{ USUBJID }{Unique Subject Identifier}
\item{ ISSEQ }{Sequence Number}
\item{ ISTESTCD }{Immunogenicity Test/Exam Short Name}
\item{ ISTEST }{Immunogenicity Test or Examination Name}
\item{ ISCAT }{Category for Immunogenicity Test}
\item{ ISORRES }{Results or Findings in Original Units}
\item{ ISORRESU }{Original Units}
\item{ ISSTRESC }{Character Result/Finding in Std Format}
\item{ ISSTRESN }{Numeric Results/Findings in Std. Units}
\item{ ISSTRESU }{Standard Units}
\item{ ISSTAT }{Completion Status}
\item{ ISREASND }{Reason Not Done}
\item{ ISNAM }{Vendor Name}
\item{ ISSPEC }{Specimen Type}
\item{ ISMETHOD }{Method of Test or Examination}
\item{ ISBLFL }{Baseline Flag}
\item{ ISLLOQ }{Lower Limit of Quantitation}
\item{ VISITNUM }{Visit Number}
\item{ EPOCH }{Epoch}
\item{ ISDTC }{Date/Time of Collection}
\item{ ISDY }{Study Day of Visit/Collection/Exam}
\item{ ISULOQ }{Upper Limit of Quantitation}
\item{ IDVARVAL }{Identifying Variable Value}
\item{ LOD }{Limit of Detection}
\item{ AVISITN }{Analysis Visit (N)}
\item{ AVISIT }{Analysis Visit}
\item{ ATPTN }{Analysis Timepoint (N)}
\item{ ATPT }{Analysis Timepoint}
\item{ ATPTREF }{Analysis Timepoint Reference}
\item{ ADT }{Analysis Date}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ PPROTFL }{Per-Protocol Population Flag}
\item{ ADY }{Analysis Relative Day}
\item{ DERIVED }{Derivation Method}
\item{ PARAMCD }{Parameter Code}
\item{ PARAM }{Parameter}
\item{ PARAMN }{Parameter (N)}
\item{ PARCAT1 }{Parameter Category 1}
\item{ CUTOFF02 }{First Cutoff Value}
\item{ CUTOFF03 }{Second Cutoff Value}
\item{ AVAL }{Analysis Value}
\item{ AVALU }{Analysis Value Unit}
\item{ SERCAT1 }{Pre-vaccination seropositivity status}
\item{ SERCAT1N }{Pre-vaccination sero status (n)}
\item{ DTYPE }{Derivation Type}
\item{ ABLFL }{Baseline Record Flag}
\item{ BASE }{Baseline Value}
\item{ BASETYPE }{Baseline Type}
\item{ BASECAT1 }{Baseline Category 1}
\item{ CHG }{Change from Baseline}
\item{ R2BASE }{Ratio to Baseline}
\item{ CRIT1FL }{Criterion 1 Evaluation Result Flag}
\item{ CRIT1 }{Analysis Criterion 1}
\item{ CRIT1FN }{Criterion 1 Evaluation Result Flag (N)}
\item{ APERIOD }{Period}
\item{ APERSDT }{Period Start Date}
\item{ APEREDT }{Period End Date}
\item{ TRTA }{Actual Treatment}
\item{ TRTP }{Planned Treatment}
\item{ PPSRFL }{Per-Protocol Record-Level Flag}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ SITEID }{Study Site Identifier}
\item{ INVID }{Investigator Identifier}
\item{ INVNAM }{Investigator Name}
\item{ BRTHDTC }{Date/Time of Birth}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ ETHNIC }{Ethnicity}
\item{ ARMCD }{Planned Arm Code}
\item{ ARM }{Description of Planned Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ COUNTRY }{Country}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT02P }{Planned Treatment for Period 02}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRT02A }{Actual Treatment for Period 02}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ SAFFL }{Safety Population Flag}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ REGION1 }{Geographic Region 1}
\item{ VAX01DT }{Vaccination Date 01}
\item{ VAX02DT }{Vaccination Date 02}
\item{ AP01SDT }{Period 01 Start Date}
\item{ AP01EDT }{Period 01 End Date}
\item{ AP02SDT }{Period 02 Start Date}
\item{ AP02EDT }{Period 02 End Date}
}
}
\source{
Generated from admiralvaccine package (template ad_adis.R).
}
\usage{
adis_vaccine
}
\description{
Immunogenicity Specimen Assessments
}
\details{
Contains a set of 16 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
I0019NLF \tab LOG10 4FOLD (I0019NT Antibody)\cr
I0019NT  \tab I0019NT Antibody              \cr
I0019NTF \tab 4FOLD (I0019NT Antibody)      \cr
I0019NTL \tab LOG10 (I0019NT Antibody)      \cr
J0033VLF \tab LOG10 4FOLD (J0033VN Antibody)\cr
J0033VN  \tab J0033VN Antibody              \cr
J0033VNF \tab 4FOLD (J0033VN Antibody)      \cr
J0033VNL \tab LOG10 (J0033VN Antibody)      \cr
M0019LLF \tab LOG10 4FOLD (M0019LN Antibody)\cr
M0019LN  \tab M0019LN Antibody              \cr
M0019LNF \tab 4FOLD (M0019LN Antibody)      \cr
M0019LNL \tab LOG10 (M0019LN Antibody)      \cr
R0003MA  \tab R0003MA Antibody              \cr
R0003MAF \tab 4FOLD (R0003MA Antibody)      \cr
R0003MAL \tab LOG10 (R0003MA Antibody)      \cr
R0003MLF \tab LOG10 4FOLD (R0003MA Antibody)
}
}
\examples{
data("adis_vaccine")
}
\references{
None
}
\keyword{datasets}
