% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlbhy.R
\docType{data}
\name{adlbhy}
\alias{adlbhy}
\title{Dataset adlbhy}
\format{
A data frame with 14 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ TRT01A }{TRT01A}
\item{ PARAMCD }{PARAMCD}
\item{ LBSEQ }{LBSEQ}
\item{ ADT }{ADT}
\item{ AVISIT }{AVISIT}
\item{ ADY }{ADY}
\item{ AVAL }{AVAL}
\item{ ANRHI }{ANRHI}
\item{ CRIT1 }{CRIT1}
\item{ CRIT1FL }{CRIT1FL}
\item{ AVALC }{AVALC}
\item{ PARAM }{PARAM}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adlbhy
}
\description{
adlbhy dataset
}
\examples{
data("adlbhy")
}
\references{
None
}
\keyword{datasets}
