% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpp.R
\docType{data}
\name{adpp}
\alias{adpp}
\title{Dataset adpp}
\format{
A data frame with 80 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ PPTESTCD }{PPTESTCD}
\item{ PPTEST }{PPTEST}
\item{ PPCAT }{PPCAT}
\item{ PPORRES }{PPORRES}
\item{ PPORRESU }{PPORRESU}
\item{ PPSTRESU }{PPSTRESU}
\item{ PPSPEC }{PPSPEC}
\item{ PPRFDTC }{PPRFDTC}
\item{ TRTSDT.x }{TRTSDT.x}
\item{ TRTEDT.x }{TRTEDT.x}
\item{ DTHDT.x }{DTHDT.x}
\item{ EOSDT.x }{EOSDT.x}
\item{ TRT01P.x }{TRT01P.x}
\item{ TRT01A.x }{TRT01A.x}
\item{ ADT }{ADT}
\item{ ADY }{ADY}
\item{ PARAMCD }{PARAMCD}
\item{ AVAL }{AVAL}
\item{ AVALC }{AVALC}
\item{ SRCDOM }{SRCDOM}
\item{ SRCVAR }{SRCVAR}
\item{ SRCSEQ }{SRCSEQ}
\item{ VISIT }{VISIT}
\item{ VISITNUM }{VISITNUM}
\item{ AVISIT }{AVISIT}
\item{ AVISITN }{AVISITN}
\item{ TRTP }{TRTP}
\item{ TRTA }{TRTA}
\item{ AVALCA1N }{AVALCA1N}
\item{ AVALCAT1 }{AVALCAT1}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P.y }{TRT01P.y}
\item{ TRT01A.y }{TRT01A.y}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTSDT.y }{TRTSDT.y}
\item{ TRTEDT.y }{TRTEDT.y}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT.y }{EOSDT.y}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT.y }{DTHDT.y}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
adpp
}
\description{
adpp dataset
}
\examples{
load("./data/adpp.rda")
}
\references{
None
}
\keyword{datasets}
