% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrs_onco.R
\docType{data}
\name{adrs_onco}
\alias{adrs_onco}
\title{Tumor Response Analysis}
\format{
A data frame with 79 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ SUBJID }{Subject Identifier for the Study}
\item{ SITEID }{Study Site Identifier}
\item{ COUNTRY }{Country}
\item{ DOMAIN }{Domain Abbreviation}
\item{ RFSTDTC }{Subject Reference Start Date/Time}
\item{ RFENDTC }{Subject Reference End Date/Time}
\item{ RFXSTDTC }{Date/Time of First Study Treatment}
\item{ RFXENDTC }{Date/Time of Last Study Treatment}
\item{ RFPENDTC }{Date/Time of End of Participation}
\item{ SCRFDT }{Screen Failure Date}
\item{ FRVDT }{Final Retrieval Visit Date}
\item{ DTHDTC }{Date/Time of Death}
\item{ DTHADY }{Relative Day of Death}
\item{ DTHFL }{Subject Death Flag}
\item{ LDDTHELD }{Elapsed Days from Last Dose to Death}
\item{ LDDTHGR1 }{Last Dose to Death - Days Elapsed Grp 1}
\item{ DTH30FL }{Death Within 30 Days of Last Trt Flag}
\item{ DTHA30FL }{Death After 30 Days from Last Trt Flag}
\item{ DTHDOM }{Domain for Date of Death Collection}
\item{ DTHB30FL }{Death Within 30 Days of First Trt Flag}
\item{ ASEQ }{Analysis Sequence Number}
\item{ REGION1 }{Geographic Region 1}
\item{ DMDTC }{Date/Time of Collection}
\item{ DMDY }{Study Day of Collection}
\item{ AGE }{Age}
\item{ AGEU }{Age Units}
\item{ AGEGR1 }{Pooled Age Group 1}
\item{ SEX }{Sex}
\item{ RACE }{Race}
\item{ RACEGR1 }{Pooled Race Group 1}
\item{ ETHNIC }{Ethnicity}
\item{ SAFFL }{Safety Population Flag}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ TRT01P }{Planned Treatment for Period 01}
\item{ TRT01A }{Actual Treatment for Period 01}
\item{ TRTSDT }{Date of First Exposure to Treatment}
\item{ TRTSDTM }{Datetime of First Exposure to Treatment}
\item{ TRTSTMF }{Time of First Exposure Imput. Flag}
\item{ TRTEDT }{Date of Last Exposure to Treatment}
\item{ TRTEDTM }{Datetime of Last Exposure to Treatment}
\item{ TRTETMF }{Time of Last Exposure Imput. Flag}
\item{ EOSSTT }{End of Study Status}
\item{ EOSDT }{End of Study Date}
\item{ RFICDTC }{Date/Time of Informed Consent}
\item{ RANDDT }{Date of Randomization}
\item{ LSTALVDT }{Date Last Known Alive}
\item{ TRTDURD }{Total Treatment Duration (Days)}
\item{ DTHDT }{Date of Death}
\item{ DTHCAUS }{Cause of Death}
\item{ DTHCGR1 }{Cause of Death Reason 1}
\item{ ADT }{Analysis Date}
\item{ ADTF }{Analysis Date Imputation Flag}
\item{ AVISIT }{Analysis Visit}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ PARCAT1 }{Parameter Category 1}
\item{ PARCAT2 }{Parameter Category 2}
\item{ PARCAT3 }{Parameter Category 3}
\item{ AVAL }{Analysis Value}
\item{ AVALC }{Analysis Value (C)}
\item{ ANL01FL }{Analysis Flag 01}
\item{ ANL02FL }{Analysis Flag 02}
\item{ VISITNUM }{Visit Number}
\item{ VISIT }{Visit Name}
\item{ RSTESTCD }{Assessment Short Name}
\item{ RSTEST }{Assessment Name}
\item{ RSORRES }{Result or Finding in Original Units}
\item{ RSSTRESC }{Character Result/Finding in Std Format}
\item{ RSEVAL }{Evaluator}
\item{ RSEVALID }{Evaluator Identifier}
\item{ RSACPTFL }{Accepted Record Flag}
\item{ RSDTC }{Date/Time of Assessment}
\item{ RSSEQ }{Sequence Number}
\item{ DTHDTF }{Date of Death Imputation Flag}
}
}
\source{
Generated from admiralonco package (template ad_adrs.R).
}
\usage{
adrs_onco
}
\description{
Tumor Response Analysis
}
\details{
Contains a set of 13 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
BCP   \tab Best Overall Response of CR/PR by Investigator (confirmation not required)\cr
BOR   \tab Best Overall Response by Investigator (confirmation not required)         \cr
CB    \tab Clinical Benefit by Investigator (confirmation for response not required) \cr
CBCP  \tab Best Confirmed Overall Response of CR/PR by Investigator                  \cr
CBOR  \tab Best Confirmed Overall Response by Investigator                           \cr
CCB   \tab Confirmed Clinical Benefit by Investigator                                \cr
CRSP  \tab Confirmed Response by Investigator                                        \cr
DEATH \tab Death                                                                     \cr
LSTA  \tab Last Disease Assessment by Investigator                                   \cr
MDIS  \tab Measurable Disease at Baseline by Investigator                            \cr
OVR   \tab Overall Response by Investigator                                          \cr
PD    \tab Disease Progression by Investigator                                       \cr
RSP   \tab Response by Investigator (confirmation not required)
}
}
\examples{
data("adrs_onco")
}
\references{
None
}
\keyword{datasets}
