% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_vaccine.R
\docType{data}
\name{adsl_vaccine}
\alias{adsl_vaccine}
\title{Dataset adsl_vaccine}
\format{
A data frame with 46 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ USUBJID }{USUBJID}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ INVID }{INVID}
\item{ INVNAM }{INVNAM}
\item{ BRTHDTC }{BRTHDTC}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRT01P }{TRT01P}
\item{ TRT02P }{TRT02P}
\item{ TRT01A }{TRT01A}
\item{ TRT02A }{TRT02A}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ SAFFL }{SAFFL}
\item{ PPROTFL }{PPROTFL}
\item{ RACEGR1 }{RACEGR1}
\item{ AGEGR1 }{AGEGR1}
\item{ REGION1 }{REGION1}
\item{ VAX01DT }{VAX01DT}
\item{ VAX02DT }{VAX02DT}
\item{ AP01SDT }{AP01SDT}
\item{ AP01EDT }{AP01EDT}
\item{ AP02SDT }{AP02SDT}
\item{ AP02EDT }{AP02EDT}
}
}
\source{
Generated from pharmaverse/admiralvaccine.
}
\usage{
adsl_vaccine
}
\description{
adsl_vaccine dataset
}
\examples{
data("adsl_vaccine")
}
\references{
None
}
\keyword{datasets}
