% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtte_onco.R
\docType{data}
\name{adtte_onco}
\alias{adtte_onco}
\title{Time to Event Analysis for Oncology}
\format{
A data frame with 20 columns:
\describe{
\item{ STUDYID }{Study Identifier}
\item{ USUBJID }{Unique Subject Identifier}
\item{ ASEQ }{Analysis Sequence Number}
\item{ AGE }{Age}
\item{ SEX }{Sex}
\item{ ARM }{Description of Planned Arm}
\item{ ARMCD }{Planned Arm Code}
\item{ ACTARM }{Description of Actual Arm}
\item{ ACTARMCD }{Actual Arm Code}
\item{ ADT }{Analysis Date}
\item{ PARAM }{Parameter}
\item{ PARAMCD }{Parameter Code}
\item{ AVAL }{Analysis Value}
\item{ STARTDT }{Time-to-Event Origin Date for Subject}
\item{ CNSR }{Censor}
\item{ EVNTDESC }{Event or Censoring Description}
\item{ CNSDTDSC }{Censor Date Description}
\item{ SRCDOM }{Source Data}
\item{ SRCVAR }{Source Variable}
\item{ SRCSEQ }{Source Sequence Number}
}
}
\source{
Generated from admiralonco package (template ad_adtte.R).
}
\usage{
adtte_onco
}
\description{
Time to Event Analysis for Oncology
}
\details{
Contains a set of 3 unique Parameter Codes and Parameters: \tabular{ll}{
\strong{PARAMCD} \tab \strong{PARAM} \cr
OS  \tab Overall Survival         \cr
PFS \tab Progression Free Survival\cr
RSD \tab Duration of Response
}
}
\examples{
data("adtte_onco")
}
\references{
None
}
\keyword{datasets}
