% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advs.R
\docType{data}
\name{advs}
\alias{advs}
\title{Dataset advs}
\format{
A data frame with 103 columns:
\describe{
\item{ STUDYID }{STUDYID}
\item{ DOMAIN }{DOMAIN}
\item{ USUBJID }{USUBJID}
\item{ VSSEQ }{VSSEQ}
\item{ VSTESTCD }{VSTESTCD}
\item{ VSTEST }{VSTEST}
\item{ VSPOS }{VSPOS}
\item{ VSORRES }{VSORRES}
\item{ VSORRESU }{VSORRESU}
\item{ VSSTRESC }{VSSTRESC}
\item{ VSSTRESN }{VSSTRESN}
\item{ VSSTRESU }{VSSTRESU}
\item{ VSSTAT }{VSSTAT}
\item{ VSLOC }{VSLOC}
\item{ VSBLFL }{VSBLFL}
\item{ VISITNUM }{VISITNUM}
\item{ VISIT }{VISIT}
\item{ VISITDY }{VISITDY}
\item{ VSDTC }{VSDTC}
\item{ VSDY }{VSDY}
\item{ VSTPT }{VSTPT}
\item{ VSTPTNUM }{VSTPTNUM}
\item{ VSELTM }{VSELTM}
\item{ VSTPTREF }{VSTPTREF}
\item{ TRTSDT }{TRTSDT}
\item{ TRTEDT }{TRTEDT}
\item{ TRT01A }{TRT01A}
\item{ TRT01P }{TRT01P}
\item{ ADT }{ADT}
\item{ ADY }{ADY}
\item{ PARAMCD }{PARAMCD}
\item{ AVAL }{AVAL}
\item{ AVALC }{AVALC}
\item{ ATPTN }{ATPTN}
\item{ ATPT }{ATPT}
\item{ AVISIT }{AVISIT}
\item{ AVISITN }{AVISITN}
\item{ DTYPE }{DTYPE}
\item{ ONTRTFL }{ONTRTFL}
\item{ ANRLO }{ANRLO}
\item{ ANRHI }{ANRHI}
\item{ A1LO }{A1LO}
\item{ A1HI }{A1HI}
\item{ ANRIND }{ANRIND}
\item{ BASETYPE }{BASETYPE}
\item{ ABLFL }{ABLFL}
\item{ BASE }{BASE}
\item{ BASEC }{BASEC}
\item{ BNRIND }{BNRIND}
\item{ CHG }{CHG}
\item{ PCHG }{PCHG}
\item{ ANL01FL }{ANL01FL}
\item{ TRTP }{TRTP}
\item{ TRTA }{TRTA}
\item{ ASEQ }{ASEQ}
\item{ AVALCA1N }{AVALCA1N}
\item{ AVALCAT1 }{AVALCAT1}
\item{ PARAM }{PARAM}
\item{ PARAMN }{PARAMN}
\item{ SUBJID }{SUBJID}
\item{ RFSTDTC }{RFSTDTC}
\item{ RFENDTC }{RFENDTC}
\item{ RFXSTDTC }{RFXSTDTC}
\item{ RFXENDTC }{RFXENDTC}
\item{ RFICDTC }{RFICDTC}
\item{ RFPENDTC }{RFPENDTC}
\item{ DTHDTC }{DTHDTC}
\item{ DTHFL }{DTHFL}
\item{ SITEID }{SITEID}
\item{ AGE }{AGE}
\item{ AGEU }{AGEU}
\item{ SEX }{SEX}
\item{ RACE }{RACE}
\item{ ETHNIC }{ETHNIC}
\item{ ARMCD }{ARMCD}
\item{ ARM }{ARM}
\item{ ACTARMCD }{ACTARMCD}
\item{ ACTARM }{ACTARM}
\item{ COUNTRY }{COUNTRY}
\item{ DMDTC }{DMDTC}
\item{ DMDY }{DMDY}
\item{ TRTSDTM }{TRTSDTM}
\item{ TRTSTMF }{TRTSTMF}
\item{ TRTEDTM }{TRTEDTM}
\item{ TRTETMF }{TRTETMF}
\item{ TRTDURD }{TRTDURD}
\item{ SCRFDT }{SCRFDT}
\item{ EOSDT }{EOSDT}
\item{ EOSSTT }{EOSSTT}
\item{ FRVDT }{FRVDT}
\item{ RANDDT }{RANDDT}
\item{ DTHDT }{DTHDT}
\item{ DTHADY }{DTHADY}
\item{ LDDTHELD }{LDDTHELD}
\item{ LSTALVDT }{LSTALVDT}
\item{ AGEGR1 }{AGEGR1}
\item{ SAFFL }{SAFFL}
\item{ RACEGR1 }{RACEGR1}
\item{ REGION1 }{REGION1}
\item{ LDDTHGR1 }{LDDTHGR1}
\item{ DTH30FL }{DTH30FL}
\item{ DTHA30FL }{DTHA30FL}
\item{ DTHB30FL }{DTHB30FL}
}
}
\source{
Generated from pharmaverse/admiral.
}
\usage{
advs
}
\description{
advs dataset
}
\examples{
load("advs.rda")
}
\references{
None
}
\keyword{datasets}
