% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverseadam-package.R
\docType{package}
\name{pharmaverseadam-package}
\alias{pharmaverseadam}
\alias{pharmaverseadam-package}
\title{pharmaverseadam: ADaM Test Data for the Pharmaverse Family of Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Analysis Data Model (ADaM) datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/pharmaverseadam}
  \item \url{https://github.com/pharmaverse/pharmaverseadam}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com}

}
\concept{internal}
\keyword{internal}
