% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverseadam-package.R
\docType{package}
\name{pharmaverseadam-package}
\alias{pharmaverseadam}
\alias{pharmaverseadam-package}
\title{pharmaverseadam: ADaM Test Data for the 'Pharmaverse' Family of Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Analysis Data Model (ADaM) datasets constructed using the Study Data Tabulation Model (SDTM) datasets contained in the 'pharmaversesdtm' package and the template scripts from the 'admiral' family of packages. ADaM dataset specifications are described in the CDISC ADaM implementation guide, accessible by creating a free account on \url{https://www.cdisc.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/pharmaverseadam/}
  \item \url{https://github.com/pharmaverse/pharmaverseadam/}
}

}
\author{
\strong{Maintainer}: Fanny Gautier \email{fanny.gautier@cytel.com} (\href{https://orcid.org/0009-0004-3581-0131}{ORCID})

Authors:
\itemize{
  \item Stefan Bundfuss (\href{https://orcid.org/0009-0005-0027-1198}{ORCID})
  \item Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Lina Patil \email{lina.patil@cytel.com}
  \item Gerardo Jose Rodriguez \email{gerardo.jrac@gmail.com} (\href{https://orcid.org/0000-0003-1413-0060}{ORCID})
  \item Vladyslav Shuliar (\href{https://orcid.org/0009-0008-2354-8999}{ORCID})
}

Other contributors:
\itemize{
  \item Cytel Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
