% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverseadam-package.R
\docType{package}
\name{pharmaverseadam-package}
\alias{pharmaverseadam}
\alias{pharmaverseadam-package}
\title{pharmaverseadam: ADaM Test Data for the 'Pharmaverse' Family of Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Analysis Data Model (ADaM) datasets constructed using the Study Data Tabulation Model (SDTM) datasets contained in the 'pharmaversesdtm' package and the template scripts from the 'admiral' family of packages. ADaM dataset specifications are described in: CDISC Analysis Data Model Team (2021) \url{https://www.cdisc.org/system/files/members/standard/foundational/ADaMIG_v1.3.pdf}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/pharmaverseadam/main/}
  \item \url{https://github.com/pharmaverse/pharmaverseadam/}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com}

Authors:
\itemize{
  \item Kangjie Zhang
  \item Stefan Bundfuss
  \item Gayatri G
  \item Daphne Grassely
  \item Zelos Zhu
  \item Sadchla Mascary
}

}
\concept{internal}
\keyword{internal}
